+++ components/Tag/index.js
@@ -1,0 +1,6 @@
+import Tag from './tag.vue';
+
+Tag.install = (Vue) => {
+    Vue.component(Tag.name, Tag);
+};
+export default Tag;